Clear the read message where appropriate to avoid ASSERTing all over the place.
